Fixed model Added extension for float

+++ Expenses/Expense.swift
@@ -12,10 +12,10 @@ struct Expense {
     var date: Date
     var category: Category
     var account: Account
-    var amount: Double
+    var amount: Float
     var comment: String
     
-    init(date: Date, category: Category, account: Account, amount: Double, comment: String) {
+    init(date: Date, category: Category, account: Account, amount: Float, comment: String) {
         self.date = date
         self.category = category
         self.account = account
